MPT2-main/Backend/requirements.txt
supabase
uvicorn
fastapi
python-multipart
pydantic[email]
pydantic
python-dotenv
python-jose
mysql-connector-python
psycopg